refactor: export pacakge path
