+++ bin/index.js
@@ -1,6 +1,4 @@
-import { Bem } from "../bem-game/bem";
-import { BaseBall } from "../number-baseball/baseball";
-import { RockScissorsPaper } from "../rock-scissors-paper/rock-scissors-paper";
-import { TextAdventures } from "../text-adventures";
-
-export { Bem, BaseBall, RockScissorsPaper, TextAdventures };
+export { TextAdventures } from "./text-adventures";
+export { Bem } from "./bem-game/bem";
+export { BaseBall } from "./number-baseball/baseball";
+export { RockScissorsPaper } from "./rock-scissors-paper/rock-scissors-paper";
